requirements.txt
aiogram==3.13.1
aiohttp==3.10.5
requests==2.32.3
python-dotenv==1.0.1
peewee==3.17.6
aiogram-calendar==0.5.0